Add utils/bft to export bft specific functions for protocol-specs
